pyproject.toml
[build-system]
requires = [
    "setuptools>=64",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "iblrig"
description = "The International Brain Laboratory's decision making task"
dynamic = ["version", "readme"]
keywords = ["IBL", "neuro-science"]
requires-python = "~=3.10"
license = {file = "LICENSE"}
dependencies = [
    "colorlog",
    "cython",
    "graphviz",
    "ibllib@git+https://github.com/int-brain-lab/ibllib.git",
    "ipython",
    "iblpybpod@git+https://github.com/int-brain-lab/iblpybpod.git@no-gui",
    "iblscripts@git+https://github.com/int-brain-lab/iblscripts.git@iblrigv8",
    "project_extraction@git+https://github.com/int-brain-lab/project_extraction.git",
    "iblutil >= 1.7.1",
    "numpy",
    "ONE-api",
    "packaging",
    "pandas",
    "pydantic",
    "pyOpenSSL",
    "python-dateutil",
    "python-osc",
    "pywin32; sys_platform == 'win32'",
    "PySocks",
    "PyYAML",
    "scipy",
    "sounddevice",
    "types-requests",
]

[project.optional-dependencies]
DEV = [
    "coverage",
    "myst-parser",
    "pylint",
    "sphinx",
    "sphinx-autobuild",
    "sphinx_lesson",
    "sphinx_rtd_theme",
    "pre-commit"
]

[project.scripts]
viewsession = "iblrig.commands:viewsession"
transfer_data = "iblrig.commands:transfer_data_cli"
transfer_video_data = "iblrig.commands:transfer_video_data_cli"
transfer_ephys_data = "iblrig.commands:transfer_ephys_data_cli"
flush = "iblrig.commands:flush"
remove-old-sessions = "iblrig.commands:remove_local_sessions"
iblrig = "iblrig.gui.wizard:main"
upgrade_iblrig = "iblrig.version_management:upgrade"
install_spinnaker = "iblrig.camera:install_spinnaker_sdk"

[tool.setuptools.dynamic]
readme = {file = "README.md", content-type = "text/markdown"}
version = {attr = "iblrig.__version__"}

[tool.setuptools.packages]
find = {}
